--- main.c
@@ -1,5 +1,0 @@
-#include "uci.h"
-int main()
-{
-    Uci_Loop();
-}
