+++ src/bench.h
@@ -1,0 +1,7 @@
+#ifndef _BENCH_H_
+#define _BENCH_H_
+
+#include "uci.h"
+
+void bench(int argc, char **argv);
+#endif
